Added webhook controller and back end api and database handling. Fixed broken build.

+++ src/api/Fig.Datalayer/BusinessEntities/MemoryUsageBusinessEntity.cs
@@ -5,7 +5,7 @@ public class MemoryUsageBusinessEntity
 {
     public virtual Guid Id { get; set; }
 
-    public virtual int ClientRunTimeSeconds { get; set; }
+    public virtual double ClientRunTimeSeconds { get; set; }
     
     public virtual long MemoryUsageBytes { get; set; }
 }
